requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community==0.0.27
langchain-openai==0.0.8
langchain-core==0.1.27
openai==1.14.0
python-dotenv==1.0.1
chromadb==0.4.24
pypdf==4.0.1
docx2txt==0.8
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==1.10.13 